Re-organize sample and clear memory leaks
