+++ app/src/main/res/layout/activity_chat.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
 
-    tools:context=".ChatActivity">
+    tools:context=".database.ChatActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/messagesList"
